data-analysis/requirements.txt
Flask
flask-cors
requests
python-dotenv
vaderSentiment
tweepy
pytrends
